+++ Runtime/Extensions/ColorExtensions.cs
@@ -10,12 +10,12 @@ namespace NekoLib.Extensions
             return new(color.r, color.g, color.b, Mathf.Clamp01(alpha));
         }
 
-        public static string ToHex(this Color color)
+        public static string ColorToHex(this Color color)
         {
             return $"#{ColorUtility.ToHtmlStringRGBA(color)}";
         }
 
-        public static Color FromHex(this string hex)
+        public static Color HexToColor(this string hex)
         {
             if (ColorUtility.TryParseHtmlString(hex, out Color color))
             {