+++ Runtime/Utilities/ColorPalette/Palette.cs
@@ -1,0 +1,197 @@
+using UnityEngine;
+
+namespace NekoLib.ColorPalette
+{
+    public static class Palette
+    {
+        #region Gray Colors
+        public static Color DarkGray
+        {
+            get
+            {
+                return new Color(0.478f, 0.478f, 0.478f, 1.0f);
+            }
+        }
+
+        public static Color MediumGray
+        {
+            get
+            {
+                return new Color(0.722f, 0.722f, 0.722f, 1.0f);
+            }
+        }
+
+        public static Color PaleGray
+        {
+            get
+            {
+                return new Color(0.878f, 0.878f, 0.878f, 1.0f);
+            }
+        }
+        #endregion
+
+        #region Yellow Colors
+        public static Color GoldenAmber
+        {
+            get
+            {
+                return new Color(1.0f, 0.757f, 0.027f, 1.0f);
+            }
+        }
+
+        public static Color SunnyGold
+        {
+            get
+            {
+                return new Color(1.0f, 0.878f, 0.400f, 1.0f);
+            }
+        }
+
+        public static Color VividYellow
+        {
+            get
+            {
+                return new Color(1.0f, 0.925f, 0.251f, 1.0f);
+            }
+        }
+
+        public static Color CreamGold
+        {
+            get
+            {
+                return new Color(1.0f, 0.980f, 0.804f, 1.0f);
+            }
+        }
+        #endregion
+
+        #region Red Colors
+        public static Color CandyRed
+        {
+            get
+            {
+                return new Color(1.0f, 0.620f, 0.620f, 1.0f);
+            }
+        }
+
+        public static Color Coral
+        {
+            get
+            {
+                return new Color(1.0f, 0.420f, 0.420f, 1.0f);
+            }
+        }
+
+        public static Color VibrantRed
+        {
+            get
+            {
+                return new Color(1.0f, 0.310f, 0.310f, 1.0f);
+            }
+        }
+        #endregion
+
+        #region Green Colors
+        public static Color DeepEmerald
+        {
+            get
+            {
+                return new Color(0.290f, 0.686f, 0.314f, 1.0f);
+            }
+        }
+
+        public static Color LightEmerald
+        {
+            get
+            {
+                return new Color(0.506f, 0.780f, 0.518f, 1.0f);
+            }
+        }
+
+        public static Color MintEmerald
+        {
+            get
+            {
+                return new Color(0.647f, 0.839f, 0.655f, 1.0f);
+            }
+        }
+        #endregion
+
+        #region Blue Colors
+        public static Color AzureTeal
+        {
+            get
+            {
+                return new Color(0.0f, 0.592f, 0.655f, 1.0f);
+            }
+        }
+
+        public static Color VibrantCyan
+        {
+            get
+            {
+                return new Color(0.0f, 0.737f, 0.831f, 1.0f);
+            }
+        }
+
+        public static Color AquaCyan
+        {
+            get
+            {
+                return new Color(0.302f, 0.816f, 0.882f, 1.0f);
+            }
+        }
+        #endregion
+
+        #region Purple Colors
+        public static Color RoyalPurple
+        {
+            get
+            {
+                return new Color(0.557f, 0.141f, 0.667f, 1.0f);
+            }
+        }
+
+        public static Color Amethyst
+        {
+            get
+            {
+                return new Color(0.671f, 0.278f, 0.737f, 1.0f);
+            }
+        }
+
+        public static Color Lavender
+        {
+            get
+            {
+                return new Color(0.808f, 0.576f, 0.847f, 1.0f);
+            }
+        }
+        #endregion
+
+        #region Orange Colors
+        public static Color BurntOrange
+        {
+            get
+            {
+                return new Color(0.902f, 0.318f, 0.0f, 1.0f);
+            }
+        }
+
+        public static Color PumpkinOrange
+        {
+            get
+            {
+                return new Color(0.961f, 0.486f, 0.0f, 1.0f);
+            }
+        }
+
+        public static Color LightOrange
+        {
+            get
+            {
+                return new Color(1.0f, 0.800f, 0.502f, 1.0f);
+            }
+        }
+        #endregion
+    }
+}