+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+## [1.0.2] - 2025-06-07
+
+### Text Formatter
+
+- Improve text colorization
+- Add text formatting extensions (bold, italic, and underline)
+- Add my own color palette
+- Detect mouse over game object (2D and 3D)
+
 ## [1.0.1] - 2025-05-31
 
 ### Network Manager and Improvements